refactor: API tweaks
