+++ v3/src/components/case-table/collection-table-model.ts
@@ -27,10 +27,6 @@ export class CollectionTableModel {
     makeObservable(this)
   }
 
-  hasTrailingScrollCount(syncScrollCount: number) {
-    return this.scrollCount < syncScrollCount
-  }
-
   @action syncScrollCount(syncScrollCount: number) {
     this.scrollCount = syncScrollCount
   }
